fix: adjust betting logic for Decimal values

+++ app/schemas/bet.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel, Field, ConfigDict
 from typing import Optional
 from app.models.enum.bet_enum import BetPrediction, BetResult, BetStatus
@@ -7,14 +8,14 @@ from app.models.enum.bet_enum import BetPrediction, BetResult, BetStatus
 class BetCreate(BaseModel):
     match_id: int
     prediction: BetPrediction
-    points_bet: int = Field(ge=1) # maior-igual a 1
+    points_bet: Decimal = Field(ge=1) # maior-igual a 1
 
 class BetResponse(BaseModel):
     id: uuid.UUID
     user_id: uuid.UUID
     match_id: int
     prediction: BetPrediction
-    points_bet: int
+    points_bet: Decimal
     odds: float
     result: Optional[BetResult] = None
     status: BetStatus
